requirements.txt
# Core Deep Learning
torch==2.5.1
torchvision==0.20.1

# Data Processing
numpy==1.26.0
opencv-python==4.10.0.84

# Configuration
pyyaml==6.0.1

# Progress Bars & CLI
tqdm==4.66.5

# Optional but recommended
matplotlib==3.9.0
pillow==10.4.0
scikit-image==0.24.0

# Development (optional)
# tensorboard==2.17.0
# jupyter==1.0.0
# ipython==8.27.0 